BGP: Fix warning message when interface has IPv4 address for unnumbered
Ticket: CM-6192
Reviewed By: Donald, Daniel, Vipin
Testing Done:

In the presence of an IPv4 address that is not a /30 or /31 and the user
is trying to use BGP unnumbered, a warning message is printed which
seems confusing to the user. Fixing the warning message to indicate that
its only a v4 session isn't coming up makes it clearer.

Signed-off-by: Dinesh G Dutt <ddutt@cumulusnetworks.com>
Acked-by: Donald Sharp <sharpd@cumulusnetworks.com>
Acked-by: Daniel Walton <dwalton@cumulusnetworks.com>
Acked-by: Vipin Kumar <vipin@cumulusnetworks.com>