quagga netlink buffer size increase
Ticket: CM-7380
Reviewed-by: CCR-3478
Testing: See Bug

This is a fix to allow the netlink buffer size to grow to a reasonable number
over the current 4mb choosen.  This number was choosen by running some
smoke tests that would fail and experimenting to find a number that worked
with those tests and then adding some headroom to that number.

This is a bit of a temporary fix due to timelines instead of the real fix
which is to write the code to properly recover from a netlink buffer overflow.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>