debian: Remove unnecessary dependency on cl-utilties
The cl-utilities dependency were causing issues in two situations:

A) The cl-utilities package name has been changed but the quagga
cmaster branch was being built on two different branches, one
with the old name, one with the new name

B) People installing quagga on non-cumulus switches were experiencing
issues due to cl-utilities not being installed.  This was especially
true if they built quagga from our source code.  We only need
cl-utilities for the startt-stop-daemon wrapper so that we could
have jdoo watch watchquagga.  This is not a big deal if people are
missing this.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>