debian: Add the creation of the quagga user to quaggavty group
The creation of the quagga user was not the only place
to add the quagga user to the quaggavty group.  If
we are reinstalling quagga over a old version of
code then we need to check to see if the quagga
user is in the quaggavty and do the right thing.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>
Reviewed-by: Dave Olson <olson@cumulusnetworks.com>